scrape/pyproject.toml
[tool.poetry]
name = "scrape"
version = "0.0.1"
description = "Content scrape and clean component of project N"
authors = ["Krutarth R"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.11.8"
fastapi = "^0.61.1"
newspaper3k = "^0.2.8"
goose3 = "^3.1.6"
justext = "^2.2.0"
htmldate = "^0.7.1"
dynaconf = "^3.1.2"
gunicorn = "^20.0.4"
clean-text = "^0.3.0"
# fuzzywuzzy = {extras = ["speedup"], version = "^0.18.0"}
# news-fetch = "^0.2.5"
pytrends = "^4.7.3"
unidecode = "^1.1.1"
autoscraper = "^1.1.9"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
pytest-cov = "^2.10.1"
mypy = "^0.790"
autoflake = "^1.4"
isort = "^5.6.4"
black = "^20.8b1"
